fix: fix Integer types

+++ src/main/java/com/example/spring/repository/UserCompanyStatusRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserCompanyStatusRepository extends JpaRepository<UserCompanyStatus, Long> {
-    UserCompanyStatus findUserCompanyStatusByUserIdAndCompanyId(String userId, Long companyId);
+public interface UserCompanyStatusRepository extends JpaRepository<UserCompanyStatus, Integer> {
+    UserCompanyStatus findUserCompanyStatusByUserIdAndCompanyId(String userId, Integer companyId);
 
-    List<UserCompanyStatus> findByUserIdAndCompanyIdIn(String userId, List<Long> companyIds);
+    List<UserCompanyStatus> findByUserIdAndCompanyIdIn(String userId, List<Integer> companyIds);
 }
